refactor: Refactor task management with TaskService class
The code changes refactor the task management by introducing the `TaskService` class. This class replaces the previous `TaskData` class and provides methods to add, update, and delete tasks. It utilizes the `TaskModel` class to create and manipulate tasks.

This refactoring improves the organization and separation of concerns in the codebase, making it easier to manage tasks in the application.
