+++ lib/widgets/tasks_list.dart
@@ -1,14 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:to_do/models/task_data.dart';
+import 'package:to_do/services/task_service.dart';
 import 'package:to_do/widgets/task_card.dart';
 
+/// A widget that displays a list of tasks.
 class TasksList extends StatelessWidget {
   const TasksList({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Consumer<TaskData>(
+    return Consumer<TaskService>(
       builder: (context, taskData, child) {
         return ListView.builder(
           itemBuilder: (context, index) {
